SQL/requirements.txt
openpyxl>=3.1.2
pandas>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
json5>=0.9.0
